Simplify admin APIs and fix controller structure
- Simplified admin get users API to return all users without pagination/filtering
- Added GetAllUsersAsync method to UserManagementService
- Split AdminCategoryManagement into separate controllers:
  * ManageCategoryController.cs for create/update/delete operations
  * GetCategoriesController.cs for getting all categories
  * GetCategoryByIdController.cs for getting category by ID
- Each controller now has single responsibility following defined pattern
